backend/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
Brotli==1.1.0
cachetools==5.5.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
cssselect2==0.8.0
faiss-cpu==1.12.0
fastapi==0.116.1
fonttools==4.59.2
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-core==2.4.3
google-cloud-documentai==3.6.0
google-cloud-storage==3.3.1
google-crc32c==1.7.1
google-genai==1.36.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.74.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.76
langchain-text-splitters==0.3.11
langsmith==0.4.27
Markdown==3.9
numpy==2.3.3
orjson==3.11.3
packaging==25.0
pillow==11.3.0
proto-plus==1.26.1
protobuf==6.32.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.11.9
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydyf==0.11.0
pypdf==6.0.0
pyphen==0.17.2
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
sniffio==1.3.1
starlette==0.47.3
tenacity==9.1.2
tinycss2==1.4.0
tinyhtml5==2.0.0
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
weasyprint==66.0
webencodings==0.5.1
websockets==15.0.1
zopfli==0.2.3.post1
zstandard==0.24.0